containers: Workaround Podman permissions with 'z' mounting
Running the pipeline with the `-profile podman` option has always failed
due to a '.command.sh: permission denied' error. As referenced in an
nf-core issue at https://github.com/nf-core/nanoseq/issues/74, this is
likely an SELinux problem on Fedora, and should be able to be worked around
by using the 'z' mounting flag. While that didn't work for the OP of that
issue, it works on my machine, so add 'z' mount flags to the working
directories and the Kraken2 database.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>