organization: Use self-contained visualization generator
The code used to generate alignment visualizations was hosted on GitHub at
https://github.com/MillironX/igv-bundler, but that code required multiple
branches to serve multiple projects and kept related code in different
locations with no real separation of concerns. Move that code to this repo,
update the git meta documents to handle the new code type, and update the
workflow script to copy files directly instead of cloning from Github.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>