versatility: Replace hard-coded genome size with dynamically-calculated one
Canu requires an approximate genome size to function. As this pipeline is
focused on Japanese Encephalitis Virus (JEV), the approximate size of the
JEV genome has been hard-coded as a magic number since 555226daff7d ("Swap
MetaVelvet for Canu"). To improve this, parse the genome size from genbank,
emit that as a value channel, and update the assembly workflow and process
to accept the size parameter.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>