generalization: Format haplotype output for multi-dimensional results
Since cb2b3284cc9d ("Switch to using type-based function calls instead of
dataframe manipulation"), the haplotype-finder script has output a TSV of
the linkage haplotype occurrences and the linkage disequilibrium
statistics. This file was formatted in a rigid 2-D structure that could not
be updated to multi-dimensional haplotype searching, so drop it from the
script, and update the pipeline to get along without it.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>