bug: Check for empty NCBI genome files
The NCBI e-utilities return exit code 0 (success) when the fail due to
too many requests, and the redirection produces a file with a single
newline. Check if the output file is only whitespace, and fail the process
if that is the case so it may be retried by the pipeline.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>