ci: Be more conservative in testing matrix combinations
While the pipeline has many parameters that need to be tested, not every
combination of every parameter needs to be tested. For example, it can be
reasonably assumed that the `--skip_...` flags do not interact with the
Nextflow version or the container engine, but they need to be tested
nonetheless. Interacting components that remain in the matrix are

- Nextflow version
    Controls container pull, and influences channel flow
- Container engine
    Pulls different images per read type, and is controlled by nf version
- Read type

The `--skip_...` flags have been relegated to `include:` entries in the
execution matrix, and using the edge release has had its own provisions
since 548813d1db74 ("ci: Only test full pipeline on edge release").

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>