performance: Dynamically allocate Kraken2 memory based on database size
The Kraken2 classification process has historically been hard to manage
thanks to the dramatic differences in memory requirements from the database
size. Automatically calculate the database size and allocate the process
memory based on that. To allow for the memory required to hold the reads
and the database, a fractional power-law term is added to the size of the
database to determine the allocation size.

Signed-off-by: Thomas A. Christensen II <25492070+MillironX@users.noreply.github.com>