Fix versioned accession numbers mismatch
seqret does not keep version numbers when converting GenBank to fasta. In
order to match the version-less accession numbers, use regular expressions
to remove any dot-digit from the accession numbers. This should allow users
to continue to use version numbers if they so choose, but not require one
way or the other.